refactor(applications): extract AIEvaluationDisplay sub-components to sibling file
Moves ScoreDisplay, DecisionDisplay, DisqualificationReason, EvaluationSummary,
ImprovementRecommendations, and StatusChip from AIEvaluationDisplay.tsx into a
new AIEvaluationDisplay.parts.tsx so the main file drops from 628 to 355 lines,
back under the oversized-file threshold. Also replaces array-index keys in the
extracted list renderers with stable content-derived keys, addressing the
react-doctor no-array-index-key/no-array-index-as-key warnings on those lines.
